feat: modified landing page
