+++ src/app/components/mint-unmint/components/landing-page/components/setup-information/components/setup-information-wallet-requirement.tsx
