+++ src/app/components/mint-unmint/components/protocol-summary-stack/components/protocol-summary-stack.layout.tsx
@@ -3,7 +3,7 @@ import { HasChildren } from '@models/has-children';
 
 export function ProtocolSummaryStackLayout({ children }: HasChildren): React.JSX.Element {
   return (
-    <HStack py={'25px'} alignContent={'start'} h={'auto'} w={'100%'} spacing={'25px'}>
+    <HStack w={'50%'} spacing={'25px'}>
       {children}
     </HStack>
   );
