+++ src/app/pages/dashboard/dashboard.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
+import { LandingPage } from '@components/mint-unmint/components/landing-page/landing-page';
 import { MintUnmint } from '@components/mint-unmint/mint-unmint';
 import { MyVaultsSmall } from '@components/my-vaults-small/my-vaults-small';
 import { PageLayout } from '@pages/components/page.layout';
@@ -11,8 +12,14 @@ export function Dashboard(): React.JSX.Element {
 
   return (
     <PageLayout>
-      <MintUnmint address={address} />
-      <MyVaultsSmall address={address} />
+      {address ? (
+        <>
+          <MintUnmint address={address} />
+          <MyVaultsSmall address={address} />
+        </>
+      ) : (
+        <LandingPage />
+      )}
     </PageLayout>
   );
 }
